requirements.txt
Flask==3.0.3
Flask-WTF==1.2.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9  # 如果用 PostgreSQL
gunicorn==21.2.0
python-dotenv==1.0.1
qrcode
requests